+++ ptolemy/data/properties/lattice/typeSystem_C/UChar.java
@@ -56,7 +56,7 @@ public class UChar extends LatticeProperty implements TypeProperty {
    }
 
    public LongToken getMinValue() {
-       return new LongToken((long)0);
+       return new LongToken(0);
    }
 
    public boolean hasMinMaxValue() {
